models のディレクトリ構成を変更。
- models.py をパッケージに変換。
- LightAssembly も models/ に含めた。
- LightAssembly と Assembly の相互変換は models/conversion/ にまとめた。
