--- src/nasap_net/light_assembly/__init__.py
@@ -1,5 +1,0 @@
-from .models import LightAssembly
-from .to_light import ConversionResult, InconsistentComponentKindError, \
-    convert_assemblies_to_light_ones
-from .to_rich import convert_light_assemblies_to_rich_ones, \
-    convert_light_assembly_to_rich_one
