+++ src/nasap_net/models/conversion/__init__.py
@@ -1,0 +1,4 @@
+from .light_to_rich import convert_light_assemblies_to_rich_ones, \
+    convert_light_assembly_to_rich_one
+from .rich_to_light import ConversionResult, InconsistentComponentKindError, \
+    convert_assemblies_to_light_ones
