+++ src/nasap_net/models/conversion/tests/test_light_to_rich.py
@@ -1,6 +1,5 @@
-from nasap_net.light_assembly import convert_light_assemblies_to_rich_ones
-from nasap_net.light_assembly.models import LightAssembly
-from nasap_net.models import Assembly, AuxEdge, Bond, Component
+from nasap_net.models import Assembly, AuxEdge, Bond, Component, LightAssembly
+from nasap_net.models.conversion import convert_light_assemblies_to_rich_ones
 
 
 def test():
