+++ src/nasap_net/models/conversion/tests/test_round_trip.py
@@ -1,7 +1,6 @@
-from nasap_net.light_assembly import convert_assemblies_to_light_ones, \
+from nasap_net.models import Assembly, AuxEdge, Bond, Component, LightAssembly
+from nasap_net.models.conversion import convert_assemblies_to_light_ones, \
     convert_light_assemblies_to_rich_ones
-from nasap_net.light_assembly.models import LightAssembly
-from nasap_net.models import Assembly, AuxEdge, Bond, Component
 
 
 def test_rich_to_light_to_rich():
